Moved PGNTerminalSymbol class to Sandra.Chess/PGN/PGNSyntax.cs, under namespace Sandra.PGN.
